+++ cli/docs/changelog.md
@@ -1,7 +1,7 @@
 # TBD
 
 ### Features
-* Added `rendertemplate` to `kurtosis files` which allows you to upload a golang text/template with data provided as a json file.
+* Added `rendertemplate` to `kurtosis files` which allows you to upload a Golang text/template with data provided as a JSON file.
 
 # 0.28.0